Project code/requirements.txt
PyYAML==6.0.1
numpy==1.26.4